pyproject.toml
[project]
name = "linkedin-mcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Gil Fernandes", email = "gil.fernandes@onepointltd.com" }
]
requires-python = ">=3.13"
dependencies = [
    "aiohttp>=3.13.2",
    "bs4>=0.0.2",
    "mcp[cli]>=1.21.0",
    "pydantic>=2.12.4",
    "pytest-asyncio>=1.3.0",
    "python-dotenv>=1.2.1",
    "selenium>=4.38.0",
    "tenacity>=9.1.2",
    "webdriver-manager>=4.0.2",
]

[project.scripts]
linkedin-mcp = "linkedin_mcp:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "black>=25.11.0",
    "pytest>=9.0.1",
]
